feat: direct dynamic import
